#56 Implementace posuvníku, jako je v f-VEPu. Odstranění circle indikátoru
