+++ app/src/main/res/layout/fragment_erp.xml
@@ -1,8 +1,8 @@
-<FrameLayout
+<LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
     <TextView
         android:id="@+id/erp_title"
@@ -14,25 +14,12 @@
     <android.support.v4.view.ViewPager
         android:id="@+id/erp_pager"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginTop="24dp"/>
+        android:layout_height="match_parent"/>
 
-    <me.relex.circleindicator.CircleIndicator
-        android:id="@+id/indicator"
-        android:layout_gravity="bottom"
+    <android.support.design.widget.TabLayout
         android:layout_width="match_parent"
-        android:layout_height="16dp"
-        android:layout_marginBottom="24dp"
-        app:ci_animator="@animator/indicator_animator"
-        app:ci_animator_reverse="@animator/indicator_animator_reverse"
-        app:ci_drawable="@drawable/black_radius_square" />
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/erp_fragment_fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/fab_run_stimulator"
-        android:layout_gravity="right|bottom"
-        android:layout_margin="5dp"/>
-
-</FrameLayout>
+        android:layout_height="10dp"
+        android:id="@+id/erp_tablayout"
+        style="@style/MyTabLayout"
+        android:layout_gravity="bottom" />
+</LinearLayout>
