+++ app/src/main/res/layout/fragment_erp_screen_1.xml
@@ -1,40 +1,52 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<LinearLayout
+<FrameLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:weightSum="1"
-    android:paddingBottom="52dp">
-
-    <ListView
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:id="@+id/erp_screen_1_listview_scheme"
-        android:choiceMode="singleChoice"
-        android:padding="5dp"
-        android:layout_weight="1"/>
-
+    android:layout_height="match_parent">
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="horizontal">
-        <Button
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:id="@+id/erp_screen_1_new_scheme"
-            android:text="@string/context_set_name"
-            style="?android:buttonBarButtonStyle"
-            android:layout_gravity="left"
+        android:orientation="vertical"
+        android:weightSum="1">
+
+        <ListView
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:id="@+id/erp_screen_1_listview_scheme"
+            android:choiceMode="singleChoice"
+            android:padding="5dp"
             android:layout_weight="1"/>
-        <Button
-            android:layout_width="0dp"
+
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:id="@+id/erp_screen_1_save_all_schemes"
-            android:text="@string/save_all"
-            style="?android:buttonBarButtonStyle"
-            android:layout_gravity="right"
-            android:layout_weight="1"/>
+            android:orientation="horizontal">
+            <Button
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:id="@+id/erp_screen_1_new_scheme"
+                android:text="@string/context_set_name"
+                style="?android:buttonBarButtonStyle"
+                android:layout_gravity="left"
+                android:layout_weight="1"/>
+            <Button
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:id="@+id/erp_screen_1_save_all_schemes"
+                android:text="@string/save_all"
+                style="?android:buttonBarButtonStyle"
+                android:layout_gravity="right"
+                android:layout_weight="1"/>
+        </LinearLayout>
     </LinearLayout>
-</LinearLayout>
+
+    <android.support.design.widget.FloatingActionButton
+        android:id="@+id/erp_fab"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:src="@drawable/fab_run_stimulator"
+        android:layout_gravity="right|bottom"
+        android:layout_margin="5dp"
+        android:layout_marginBottom="32dp"/>
+</FrameLayout>
